WebIDL/tests/submissions/heycam/exception-interface-object-002.html

<!DOCTYPE html>
<title>Exception interface object [[Call]] method</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.9.1.1 Exception interface object [[Call]] method
//
// When called, exception interface objects construct and return a new
// platform exception object.

test(function() {
  var e = new DOMException();
  assert_equals(Object.getPrototypeOf(e), DOMException.prototype);
}, "DOMException object returned from DOMException constructor has the right prototype");

test(function() {
  var e = DOMException();
  assert_equals(Object.getPrototypeOf(e), DOMException.prototype);
}, "DOMException object returned from calling DOMException has the right prototype");

test(function() {
  var e = new DOMException();
  assert_equals(e.message, "");
  assert_true(!e.hasOwnProperty("message"));
}, "constructing a DOMException with no arguments does not set a message property");

test(function() {
  var e = DOMException();
  assert_equals(e.message, "");
  assert_true(!e.hasOwnProperty("message"));
}, "calling DOMException with no arguments returns a DOMException object without a message property");

test(function() {
  var e = new DOMException({ toString: function() { return "test"; } });
  assert_property(e, "message", { writable: true, enumerable: false, configurable: true, value: "test" });
}, "constructing a DOMException with one argument sets its message property correctly");

test(function() {
  var e = DOMException({ toString: function() { return "test"; } });
  assert_property(e, "message", { writable: true, enumerable: false, configurable: true, value: "test" });
}, "calling DOMException with one argument sets its message property correctly");

test(function() {
  var e = new DOMException("test", "extra");
  assert_property(e, "message", { writable: true, enumerable: false, configurable: true, value: "test" });
}, "additional argument passed to DOMException constructor is ignored");

test(function() {
  var e = DOMException("test", "extra");
  assert_property(e, "message", { writable: true, enumerable: false, configurable: true, value: "test" });
}, "additional argument passed when calling DOMException is ignored");
</script>
